CNN_MNIST/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1702571698061/work
cachetools==5.3.3
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1707022139797/work/certifi
charset-normalizer==3.3.2
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
cycler==0.11.0
debugpy @ file:///D:/bld/debugpy_1660619187469/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
fonttools==4.38.0
gast==0.3.3
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.2
h5py==2.10.0
idna==3.7
imageio==2.31.2
importlib-metadata==6.7.0
ipykernel @ file:///D:/bld/ipykernel_1666723449132/work
ipython @ file:///D:/bld/ipython_1651240762678/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1673615989977/work
jupyter_core @ file:///D:/bld/jupyter_core_1658332454038/work
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==2.6.3
numpy==1.18.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.0
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.5.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1702399386289/work
protobuf==3.20.3
psutil @ file:///D:/bld/psutil_1653089401496/work
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1700607939962/work
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyWavelets==1.3.0
pywin32==303
pyzmq @ file:///D:/bld/pyzmq_1656183658660/work
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
saliency==0.2.1
scikit-image==0.19.3
scipy==1.4.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.3.0
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.0
tensorflow-gpu-estimator==2.3.0
termcolor==2.3.0
tifffile==2021.11.2
torch==1.13.1
torchvision==0.14.1
tornado @ file:///D:/bld/tornado_1656937938087/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
typing_extensions==4.7.1
urllib3==2.0.7
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1699959196938/work
Werkzeug==2.2.3
wincertstore==0.2
wrapt==1.16.0
zipp==3.15.0
